Rename inferepy to inferers
