+++ inferers/Cargo.toml
@@ -1,11 +1,15 @@
 [package]
-name = "inferepy"
+name = "inferers"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[profile.dev]
+rpath = true
+
 [dependencies]
+ctrlc = "3.4.2"
 image = "0.24.7"
 ndarray = "0.15.6"
 opencv = "0.88.6"
